+++ src/main/java/lermitage/intellij/extra/icons/enablers/IconEnablerType.java
@@ -4,5 +4,8 @@ package lermitage.intellij.extra.icons.enablers;
 
 public enum IconEnablerType {
 
-    GIT_SUBMODULE_FOLDER
+    /** Detect git submodule folders. */
+    IS_GIT_SUBMODULE_FOLDER,
+    /** Detect folders containing Angular files. */
+    IS_IN_ANGULAR_FOLDER
 }